requirements.txt
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.36
langchain-groq==0.1.2
langchain-huggingface==0.0.3
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.2

sentence-transformers==5.1.2
transformers==4.57.3
huggingface-hub==0.36.0

pinecone-client==3.2.2

pypdf==6.4.0

Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
-e .
